backend/api_gateway/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
uwsgi==2.0.20
logging==0.4.9.6
PyJWT==2.6.0 